content_POC/requirements.txt
selenium==4.27.1
python-dotenv==1.0.1
webdriver-manager==4.0.2
beautifulsoup4==4.12.3
lxml==5.3.0
